pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cyclelayer"
version = "0.1.0"
description = "Physics-informed Brayton Cycle Layer for turbofan RUL prognostics (N-CMAPSS)"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }

dependencies = [
    "torch>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "h5py>=3.10",
    "matplotlib>=3.8",
    "pandas>=2.2",
    "tensorboard>=2.16",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipykernel>=6.29",
]

[project.scripts]
cyclelayer-train = "cyclelayer.scripts.train:main"
cyclelayer-evaluate = "cyclelayer.scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
